fix/organize-files #88
