+++ combine/db.js
@@ -1,16 +1,16 @@
-const mysql = require("mysql2");
+const mysql = require('mysql2');
 
 const mydb = mysql.createConnection({
-    host: "localhost",
-    user: "root",
-    password: "1234",   //로컬 db
-    database: "insideout",  //로컬 db
-    port: 3306
+  host: 'localhost',
+  user: 'root',
+  password: '9981', //로컬 db
+  database: 'insideout', //로컬 db
+  port: 3306,
 });
 
 mydb.connect((err) => {
-    if (err) console.log(err);
-    else console.log('Connected to the database');
+  if (err) console.log(err);
+  else console.log('Connected to the database');
 });
 
 module.exports = mydb;
