+++ src/pages/SignUp.js
@@ -1,4 +1,4 @@
-import SignUpForm from '../components/SignUpForm';
+import SignUpForm from '../components/SignUp/SignUpForm';
 
 export default function SignUp() {
   return <SignUpForm></SignUpForm>;
