+++ src/styles/Intros/IntroCheckList.css
