+++ src/styles/Intros/IntroCommunity.css
