+++ src/styles/Intros/IntroDiary.css
